Split NXT helpers from example/Utils into src/utils
This will make it possible to use them in the test suites
